requirements.txt
cqplus-sdk==1.0.7
peewee==3.13.1
